* added ci.yml; removed .travis.yml
* dump - add mysqldump --column-statistics parameter detection
* update go.mod
* update go.sum
* added .golangci.yml
* lint - fix all misspell
* lint - fix all staticcheck
* lint - fix all errcheck
* lint - fix all structcheck
* lint - fix all nakedret
* lint - fix all goimports
* lint - fix all unconvert
* lint - fix all whitespace

+++ mysql/parse_binary.go
@@ -20,7 +20,7 @@ func ParseBinaryUint16(data []byte) uint16 {
 }
 
 func ParseBinaryInt24(data []byte) int32 {
-	u32 := uint32(ParseBinaryUint24(data))
+	u32 := ParseBinaryUint24(data)
 	if u32&0x00800000 != 0 {
 		u32 |= 0xFF000000
 	}
